requirements.txt
av==9.2.0 
beautifulsoup4==4.11.1
entrypoints==0.4
gdown==4.5.1
gin-config==0.5.0
h5py==3.7.0
imageio==2.21.1
imageio-ffmpeg
ipython==7.19.0
kornia==0.6.8
loguru==0.6.0
lpips==0.1.4
mediapy==1.1.0
mmcv==1.6.2
ninja==1.10.2.3
numpy==1.23.3
open3d==0.16.0
opencv-python==4.6.0.66
pandas==1.5.0
Pillow==9.2.0
plotly==5.7.0
PyMCubes==0.1.2
pyransac3d==0.6.0
PyYAML==6.0
rich==12.6.0
scipy==1.9.2
tensorboard==2.9.0
torch-fidelity==0.3.0
torchmetrics==0.10.0
torchtyping==0.1.4
tqdm==4.64.1
tyro==0.3.25
appdirs
nerfacc==0.3.5
plyfile
scikit-image
trimesh
torch_efficient_distloss
umsgpack
pyngrok
cryptography==39.0.2
omegaconf==2.2.3
segmentation-refinement
xatlas
protobuf==3.20.0
jinja2
click==8.1.7
tensorboardx
termcolor
